refactor: Hoist name aliasing one level higher
